Fix availability status for Airthings Cloud sensors
AirThings API keeps returning the latest value of a sensor even if that sensor goes offline and stops reporting data. As a result, HASS also continues to report measurements like the sensor is still functioning. This CL adds a check where sensor is marked unavailable if it misses 6 data checkins (60 minutes).